+++ src/main/java/br/com/anp/microservice/article/ArticleApplication.java
@@ -1,0 +1,13 @@
+package br.com.anp.microservice.article;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ArticleApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ArticleApplication.class, args);
+	}
+
+}
