+++ src/main/java/br/com/anp/microservice/article/core/entities/HealthData.java
@@ -1,0 +1,14 @@
+package br.com.anp.microservice.article.core.entities;
+
+import lombok.*;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder(access = AccessLevel.PUBLIC)
+public class HealthData {
+
+    private String status;
+
+    private Boolean active;
+}
