--- src/main/java/br/com/devtops/microservice/article/ArticleApplication.java
@@ -1,13 +1,0 @@
-package br.com.devtops.microservice.article;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ArticleApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ArticleApplication.class, args);
-	}
-
-}
