--- src/main/java/br/com/devtops/microservice/article/config/ValidationConfig.java
@@ -1,15 +1,0 @@
-package br.com.devtops.microservice.article.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.validation.beanvalidation.MethodValidationPostProcessor;
-
-@Configuration
-public class ValidationConfig {
-    @Bean
-    public MethodValidationPostProcessor validationPostProcessor() {
-        MethodValidationPostProcessor processor = new MethodValidationPostProcessor();
-
-        return  processor;
-    }
-}
