--- src/main/java/br/com/devtops/microservice/article/exceptions/CannotFindResourceException.java
@@ -1,16 +1,0 @@
-package br.com.devtops.microservice.article.exceptions;
-
-import br.com.devtops.microservice.article.exceptions.protocols.CustomException;
-
-public class CannotFindResourceException  extends CustomException {
-    static public String defaultName = "CANNOT_FIND_RESOURCE_EXCEPTION";
-    static public String defaultMessage = "Resource not found.";
-
-    public CannotFindResourceException(String message){
-        super(defaultName, message);
-    }
-
-    public CannotFindResourceException(){
-        super(defaultName, defaultMessage);
-    }
-}
