--- src/main/java/br/com/devtops/microservice/article/exceptions/UnexpectedException.java
@@ -1,21 +1,0 @@
-package br.com.devtops.microservice.article.exceptions;
-
-import br.com.devtops.microservice.article.exceptions.protocols.CustomException;
-import org.springframework.lang.Nullable;
-
-public class UnexpectedException extends CustomException {
-    static public String defaultName = "UNEXPECTED_EXCEPTION";
-    static public String defaultMessage = "An unexpected error occurred!";
-
-    public UnexpectedException(String message) {
-        super(defaultName, message);
-    }
-
-    public UnexpectedException() {
-        super(defaultName, defaultMessage);
-    }
-
-    public UnexpectedException(String message, @Nullable String stack) {
-        super(defaultName, message, stack);
-    }
-}
