+++ src/test/java/br/com/anp/microservice/article/ArticleApplicationTests.java
@@ -1,0 +1,7 @@
+package br.com.anp.microservice.article;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ArticleApplicationTests { }
