--- src/test/java/br/com/devtops/microservice/article/ArticleApplicationTests.java
@@ -1,7 +1,0 @@
-package br.com.devtops.microservice.article;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ArticleApplicationTests { }
